backend/requirements.txt
Flask
base58
pillow
numpy
flask_socketio
flask-cors
